[BE #29] fix : id type 수정

+++ BE/src/main/java/kr/codesquad/airbnb11/domain/review/ReviewRepository.java
@@ -2,6 +2,6 @@ package kr.codesquad.airbnb11.domain.review;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReviewRepository extends CrudRepository<Review, Long> {
+public interface ReviewRepository extends CrudRepository<Review, Integer> {
 
 }
